refactored and added click support
